Add files via upload
This is a Manchester Carry Chain Adder and Normal Adder, with and without an Inverter, written in Hspice. I used feature sizes of 45nm and 90 nm. 
The picture Voltage-Configuration shows my defined voltages. I used a frequency of 100 MHz.
The picture Final-Result shows the power consumption of these two adders with 45 and 90 nm technologies.